Move PostgresMarshall to wire-api
This way we can write instances for various types next to them
